Version 2.0
Created some event dispatchers for communication between the damage system component and unitbase, then unit base to UI elements to update health and attack that is displayed to the players screen.

Health and Attack are both fully linked between the damage system component and UWB Abilities bar. The health bar is now using the new system to update its fill meter as well.