Version 2.1
Changed the way we switch between mouse look mode to mouse cursor pointing mode. Before we had to hold RMB but this caused some issues with widget focus (turns out to be an engine issue after doing some digging around and getting help from the UE Developer forums). As a result, we now press RMB to toggle between the modes.

Added onhovered / onUnhovered event functionality to the unitbase bp class so that when the mouse cursor is hovered over them, their health bar is displayed to us; and when unhovered, the health bar is hidden.