Version 2.3
Fixed a bug that was keeping the cached unit selected even after making a multi selection; and in certain cases, the cached unit would have its selected indicators off but would still be moveable/commandable.

Fixed the spawning function of the buildingbase actor so that it only spawns units if there is space.

Fixed the camera controls so that RMB toggle and Hold shift methods do not conflict anymore.

Fixed a bug that would prevent selection of a single unit when having performed a multi selection right before.