Version 2.5
Added Asset pack to my project and placed some assets into the map.

Changed the building static mesh to a stylised building from the newly added asset pack.

Fixed a bug in which the hud class was trying to access a dead unit

Tried to implement mouse scroll wheel zoom functionality but no luck so far.